+++ src/library/DIPS.Mobile.UI/API/Builder/IDIPSUIOptions.cs
@@ -10,7 +10,7 @@ public interface IDIPSUIOptions
     /// </summary>
     /// <param name="callback">The method to invoke, receiving the <see cref="ContextMenuItem"/> that was tapped.</param>
     /// <returns></returns>
-    IDIPSUIOptions SetContextMenuItemClickedCallback(Action<ContextMenuItem> callback);
+    IDIPSUIOptions HandleContextMenuLogging(Action<ContextMenuLoggingMetadata> callback);
 
     /// <summary>
     ///     Sets whether DIPS.Mobile.UI should try and auto-resolve memory leaks that occur during page navigation and closing of bottomsheets.