+++ src/library/DIPS.Mobile.UI/Components/ContextMenus/ContextMenuLoggingMetadata.cs
@@ -1,0 +1,3 @@
+namespace DIPS.Mobile.UI.Components.ContextMenus;
+
+public record ContextMenuLoggingMetadata(ContextMenuItem ContextMenuItem, ContextMenu ContextMenu);