+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [41.0.0]
+- [ContextMenu] `SetContextMenuItemClickedCallback` is renamed to `HandleContextMenuLogging`.
+- [ContextMenu] `HandleContextMenuLogging` now sends a `ContextMenuLoggingMetadata`.
+- [ContextMenu] Added `ItemsShouldLogWhenClicked` which controls if all Context Menu item will send a logging callback. Default false.
+- [ContextMenu] Added `Mode` getter to know which mode a menu has.
+- [ContextMenuItem] Added `ShouldLogWhenClicked? which controls if it will send a logging callback. Default false.
+
 ## [40.3.9]
 - [Entry][iOS] Added 'Done' toolbar item. 
 - [Entry][iOS] Attempt to avoid seemingly random crash.